Let Claude extract ShellWindow from shell.rs
